WIP: moved to preparator usage

+++ Framework/src/main/java/de/rub/nds/x509anvil/framework/anvil/parameter/VersionParameter.java
@@ -16,8 +16,9 @@ import de.rub.nds.anvilcore.model.parameter.ParameterScope;
 import de.rub.nds.x509anvil.framework.anvil.ContextHelper;
 import de.rub.nds.x509anvil.framework.anvil.X509AnvilParameterType;
 import de.rub.nds.x509anvil.framework.x509.config.X509CertificateChainConfig;
-import de.rub.nds.x509anvil.framework.x509.config.X509CertificateConfig;
+import de.rub.nds.x509attacker.config.X509CertificateConfig;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -55,6 +56,6 @@ public class VersionParameter extends CertificateSpecificParameter<Integer> {
 
     @Override
     public void applyToCertificateConfig(X509CertificateConfig certificateConfig, DerivationScope derivationScope) {
-        certificateConfig.setVersion(getSelectedValue());
+        certificateConfig.setVersion(BigInteger.valueOf(getSelectedValue()));
     }
 }
